requirements_calcula.txt
asttokens==2.4.1
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
filelock==3.16.1
fsspec==2024.9.0
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
packaging==24.1
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
pooch==1.8.2
praat-parselmouth==0.4.5
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==6.1.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308; platform_system=="Windows"
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
six==1.16.0
soundfile==0.12.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.1
torchaudio==2.5.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.25.0
wcwidth==0.2.13
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
attrs==24.2.0
bitsandbytes==0.44.1
cryptography==43.0.3
datasets==3.1.0
dill==0.3.8
faiss-cpu==1.9.0
frozenlist==1.5.0
multidict==6.1.0
multiprocess==0.70.16
propcache==0.2.0
pyarrow==18.0.0
pycryptodome==3.21.0
sacremoses==0.1.1
sentence-transformers==3.2.1
xxhash==3.5.0
yarl==1.17.1

